Add arity-4 WARP example, fix WARP uint8 hard inference, migrate docs to Sphinx
- examples/train_mnist_warp.py: arity-n WARP (WARPNetN) on MNIST (default n=4)
- fix: WARPLayer/WARPLayerN.forward_hard cast to float (2*x-1 underflowed on
  uint8 inputs from train_model -> garbage hard circuit); +regression tests
- docs: migrate MkDocs -> Sphinx + pydata-sphinx-theme with autosummary API
  reference (the NumPy/PyTorch look); user guide + generated API; update docs.yml
- remove ImageNet helper modules with machine-specific paths from the package
  (kept locally under experiments/)

+++ silogic/warp.py
@@ -86,7 +86,9 @@ class WARPLayer(nn.Module):
 
     @torch.no_grad()
     def forward_hard(self, x):
-        # hard input select (argmax conn), binary inputs -> exact 2-LUT
+        # hard input select (argmax conn), binary inputs -> exact 2-LUT.
+        # cast to float first: x may be uint8, but _z does 2*x-1 (would underflow).
+        x = x.float()
         sa = torch.gather(self.cand_a, 1, self.conn_a.argmax(1, keepdim=True)).squeeze(1)
         sb = torch.gather(self.cand_b, 1, self.conn_b.argmax(1, keepdim=True)).squeeze(1)
         return (self._z(x[:, sa], x[:, sb]) > 0).float()
@@ -157,6 +159,7 @@ class WARPLayerN(nn.Module):
 
     @torch.no_grad()
     def forward_hard(self, x):
+        x = x.float()                                    # uint8-safe: 2*x-1 below
         sel = torch.gather(self.cand, 2, self.conn.argmax(2, keepdim=True)).squeeze(2)  # [out,n]
         u = 2 * x[:, sel] - 1                            # [B,out,n] in {-1,1}
         z = (self._monomials(u) * self.theta[None]).sum(2)
